[ooEngine] Add of Sprite object implementation - fix a warning in tinyxml2 lib - add GetCurrentTime function in tools namespace

+++ cppdev/ooEngine/ooEngine/tinyxml2.cpp
@@ -2065,7 +2065,7 @@ namespace tinyxml2
                         while( p < q ) {
                             const size_t delta = q - p;
                             // %.*s accepts type int as "precision"
-                            const int toPrint = ( INT_MAX < delta ) ? INT_MAX : delta;
+                            const int toPrint = static_cast< int >( ( INT_MAX < delta ) ? INT_MAX : delta );
                             Print( "%.*s", toPrint, p );
                             p += toPrint;
                         }
